requirements.txt
coreapi>=2.3
Django<2.1
django-crispy-forms>=1.7
django-filter>=2.0
django-guardian>=1.4
djangorestframework>=3.8
drf-yasg>=1.10
Markdown>=2.6
mysqlclient>=1.3.7
Pygments>=2.2